apps/auto-clipper-api/requirements.txt
fastapi
uvicorn
yt-dlp
faster-whisper
moviepy==1.0.3
librosa
numpy
scipy
python-multipart
sqlalchemy
google-generativeai
pyannote.audio
torch
torchaudio

opencv-python
python-dotenv

requests
Pillow
celery
redis
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
